add `ViewerStateProvider`, reorg view w/ TSX
